django_project/{{ cookiecutter.project_slug }}/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.4.1
dj-rest-auth==6.0.0
djangorestframework-simplejwt==5.3.1
